backend/requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.22.1
botocore==1.25.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.4
djongo==1.3.6
dnspython==2.2.1
fastapi==0.75.0
h11==0.13.0
idna==3.3
imageio==2.16.1
jmespath==1.0.0
mongoengine==0.24.1
multimethod==1.7
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycodestyle==2.8.0
pydantic==1.9.0
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
s3transfer==0.5.2
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.2.4
starlette==0.17.1
tangled-up-in-unicode==0.2.0
tifffile==2022.3.25
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
visions==0.7.5
